feat att edit

+++ src/components/Horario/DeleteHorario.js
@@ -2,22 +2,26 @@
 import React from 'react';
 import IconButton from "@mui/material/IconButton";
 import DeleteIcon from "@mui/icons-material/Delete";
-// ...
+import { toast } from "react-toastify";
 
-const DeleteHorario = ({ setError, toast, onDelete, setLoading, horarioData }) => {
+const DeleteHorario = ({ setError, onDelete, setLoading, horarioData }) => {
   const handleDeleteClick = async () => {
     if (!setLoading) {
       console.error("setLoading is not defined");
       return;
     }
-  
+
+    if (!horarioData || typeof horarioData !== 'object' || !('ID' in horarioData)) {
+      console.error("Invalid or missing horarioData");
+      return;
+    }
+
     setLoading(true);
-  
-    // Verifica se setError está definido antes de utilizá-lo
+
     if (setError) {
       setError(null);
     }
-  
+
     const requestOptions = {
       method: 'DELETE',
       headers: {
@@ -25,27 +29,26 @@ const DeleteHorario = ({ setError, toast, onDelete, setLoading, horarioData }) =
       },
       body: JSON.stringify({ ID: horarioData.ID }),
     };
-  
+
     try {
       const response = await fetch(
         'http://localhost:8080/Horario/deleteHorario.php',
         requestOptions
       );
-  
+
       console.log('Response status:', response.status);
       console.log('Response text:', await response.text());
-  
+
       if (response.ok) {
         toast.success('Horário excluído com sucesso!');
-        // Verifica se onDelete está definido antes de utilizá-lo
         if (onDelete) {
-          onDelete(); // Chama a função onDelete do componente pai
+          onDelete();
         }
       } else {
         console.error(
           `Erro na requisição: ${response.status} ${response.statusText}`
         );
-        // Verifica se setError está definido antes de utilizá-lo
+
         if (setError) {
           setError(
             'Erro durante a exclusão do horário. Consulte a console para detalhes.'
@@ -57,7 +60,7 @@ const DeleteHorario = ({ setError, toast, onDelete, setLoading, horarioData }) =
       }
     } catch (error) {
       console.error('Erro durante a exclusão do horário:', error.message);
-      // Verifica se setError está definido antes de utilizá-lo
+
       if (setError) {
         setError(
           'Erro durante a exclusão do horário. Consulte a console para detalhes.'
@@ -68,7 +71,6 @@ const DeleteHorario = ({ setError, toast, onDelete, setLoading, horarioData }) =
       setLoading(false);
     }
   };
-  
 
   return (
     <IconButton aria-label='delete' onClick={handleDeleteClick}>
